+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Front: Only load times when on step 2
 * Fix no distraction mode after Elementor footer update
 * Enqueue min JS file for production
+* New form field in add form: summary to display selected summary of the booking before confirming
 
 ### 1.4.2: May 19th, 2021
 * No Distraction for StormBringer theme